requirement.txt
google-generativeai==0.3.2
streamlit==1.30.0
python-dotenv==1.0.1
streamlit-option-menu==0.3.12
pillow==10.2.0
